refactor(event): move send_event function to class
Closes #253
